requirements.txt
asgiref==3.8.1
Deprecated==1.2.14
Django==5.0.7
djangorestframework==3.15.2
environs==11.0.0
img2pdf==0.5.1
lxml==5.2.2
marshmallow==3.21.3
packaging==24.1
pikepdf==9.1.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
wrapt==1.16.0
